pipeline/requirements.txt
matplotlib==3.10.3
numpy==2.3.1
pandas==2.3.0
Pillow==11.2.1
